Write output log for each tile into corresponding directory
