+++ compare-vector-tiles/package.json
@@ -9,6 +9,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "mkdirp": "^0.5.1",
     "pbf": "^1.3.5",
     "vector-tile": "^1.1.3"
   }
